+++ state-synchronizer/Cargo.toml
@@ -20,15 +20,16 @@ rand = "0.6.5"
 tokio = { version = "0.2.8", features = ["full"] }
 prometheus = { version = "0.7.0", default-features = false }
 
-libra-config = { path = "../config", version = "0.1.0" }
+channel = { path = "../common/channel", version = "0.1.0" }
 executor = { path = "../executor", version = "0.1.0" }
+libra-config = { path = "../config", version = "0.1.0" }
 libra-crypto = { path = "../crypto/crypto", version = "0.1.0" }
 libra-logger = { path = "../common/logger", version = "0.1.0" }
 libra-mempool = { path = "../mempool", version = "0.1.0"}
 libra-metrics = { path = "../common/metrics", version = "0.1.0" }
+libra-types = { path = "../types", version = "0.1.0" }
 network = { path = "../network", version = "0.1.0" }
 storage-client = { path = "../storage/storage-client", version = "0.1.0" }
-libra-types = { path = "../types", version = "0.1.0" }
 vm-runtime = { path = "../language/vm/vm-runtime", version = "0.1.0" }
 
 [dev-dependencies]
